verify_content() throws the appropriate error
